fixed network parameters mask, added corrector for existent data, fixed some tests

+++ core/src/main/java/org/bitcoinj/core/NetworkParameters.java
@@ -73,7 +73,8 @@ public abstract class NetworkParameters {
     protected Block genesisBlock;
     protected BigInteger maxTarget;
     protected int port;
-    protected long packetMagic;  // Indicates message origin network and is used to seek to the next message when stream state is unknown.
+    protected int packetMagic;  // Indicates message origin network and is used to seek to the next
+    // message when stream state is unknown.
     protected int addressHeader;
     protected int p2shHeader;
     protected int dumpedPrivateKeyHeader;
@@ -305,7 +306,7 @@ public abstract class NetworkParameters {
     }
 
     /** The header bytes that identify the start of a packet on this network. */
-    public long getPacketMagic() {
+    public int getPacketMagic() {
         return packetMagic;
     }
